==> step2_loadingtrainingdatas_vgg_transfert_modelandtraining.py <==
import numpy as np
import pylab as plt
import pandas as pd
import tensorflow as tf
import keras as keras

from keras.utils import np_utils
from keras.models import Sequential
from keras import optimizers
from keras.layers import Dense, Flatten, Conv2D, MaxPooling2D
from keras.layers import Dropout, Flatten, GlobalAveragePooling2D
from keras.optimizers import Adam
from keras.callbacks import ModelCheckpoint
from datetime import datetime

'''
PARAMETERS to change so as to improve the training
'''

#We can modify batch size and epochs to adjust improve the training
batch_size=100
epochs=50
vggsp500_learning_rate=0.1

#Use of exponential decay for learning rate
#https://keras.io/api/optimizers/learning_rate_schedules/exponential_decay/

initial_learning_rate = 0.1
vggsp500_decay_rate=0.99

lr_schedule = tf.keras.optimizers.schedules.ExponentialDecay(
    initial_learning_rate,
    decay_steps=10000,
    decay_rate=vggsp500_decay_rate,
    staircase=True)

##https://keras.io/api/losses/
vggsp500loss='categorical_crossentropy'                                 

##https://keras.io/api/optimizers/
vggsp500optimizer_name='adam'
#vggsp500optimizer=tf.keras.optimizers.Adam(learning_rate=vggsp500_learning_rate)   
vggsp500optimizer=keras.optimizers.Adam(learning_rate=vggsp500_learning_rate)   

##https://keras.io/api/metrics/
vggsp500metrics=['accuracy']                                           
       

'''
UTILITY FUNCTIONS
'''

def change_X_df__nparray_image(df_X_train_image_flattened ):
  '''
  setup_input_NN_image returns a dataframe of flaten image for x train and xtest
  then this function will change each date into a nparray list of images with 32, 32, 3 size 
  '''
  X_train_image=df_X_train_image_flattened
  nb_train=len(X_train_image.index)
  
  x_train=np.zeros((nb_train,32,32,3))
  for i in range(nb_train):
    tmp=np.array(X_train_image.iloc[i])
    tmp=tmp.reshape(32,32,3)
    x_train[i]=tmp
  return x_train

'''
PART 2 LOADING TRAINING DATAS AND FORMAT IT TO BE ACCEPTED BY THE NN
'''
#recuperation of datas 
X_train_image=pd.read_csv('datas/X_train_image.csv')
Y_train_StateClass_image=pd.read_csv('datas/Y_train_StateClass_image.csv')
Y_train_FutPredict_image=pd.read_csv('datas/Y_train_FutPredict_image.csv')


#setting up the index to Date
X_train_image=X_train_image.set_index("Date")
Y_train_StateClass_image=Y_train_StateClass_image.set_index("Date")
Y_train_FutPredict_image=Y_train_FutPredict_image.set_index("Date")

#modify dataset to np array for input to NN
x_train=change_X_df__nparray_image(X_train_image)
y_train_state=np.array(Y_train_StateClass_image)
y_train_value=np.array(Y_train_FutPredict_image)

##Setting up xtrain and ytrain
#Here we focus on predicting the future state Y_train_StateClass_image
nb_train=len(X_train_image.index)
x_train=np.zeros((nb_train,32,32,3))
for i in range(nb_train):
  tmp=np.array(X_train_image.iloc[i])
  tmp=tmp.reshape(32,32,3)
  x_train[i]=tmp
  
y_train=np.array(Y_train_StateClass_image)
#y_train=np.array(Y_train_FutPredict_image)

nb_train=len(X_train_image.index)
x_train=np.zeros((nb_train,32,32,3))
for i in range(nb_train):
  tmp=np.array(X_train_image.iloc[i])
  tmp=tmp.reshape(32,32,3)
  x_train[i]=tmp

y_train=np.array(Y_train_StateClass_image)
#y_train=np.array(Y_train_FutPredict_image)


#At this stage we have x and y to train the model
#If some 



'''
PART 3 VGGSP500 TRAINING AND SAVING
we suppose that we have loaded xtrain and ytrain
This part is based on the Design of the NN
Her we find the Vgg16 quite usefull
'''

#Importing the VGG16 model
from keras.applications.vgg16 import VGG16, preprocess_input


#In our example we need to y into categorical as it has 6 categories
nb_classes=6
y_train = np_utils.to_categorical(y_train, nb_classes)


#Loading the VGG16 model with pre-trained ImageNet weights
vgg_model = VGG16(weights='imagenet', include_top=False, input_shape=(32, 32, 3))
vgg_model.trainable = False # remove if you want to retrain vgg weights

vgg_model.summary()


##Transfert model from vgg
transfer_model = Sequential()
transfer_model.add(vgg_model)
transfer_model.add(Flatten())
transfer_model.add(Dense(128, activation='relu'))
transfer_model.add(Dropout(0.2))
transfer_model.add(Dense(6, activation='softmax'))



transfer_model.compile(loss=vggsp500loss, optimizer=vggsp500optimizer,
              metrics=vggsp500metrics)

##Saving the best model for each parameters
checkpoint = ModelCheckpoint("model/best_model"+vggsp500loss+"_"+vggsp500optimizer_name+"_Batch"+\
                             str(batch_size)+"_LR"+str(initial_learning_rate)+"_"+str(vggsp500_decay_rate)+".hdf5", \
                                monitor='loss', verbose=1, \
                                save_best_only=True, mode='auto', period=1)

##Fitting the model on the train data and labels.
history = transfer_model.fit(x_train, y_train, \
                              batch_size=batch_size, epochs=epochs, \
                              validation_split=0.2, verbose=1, shuffle=True, \
                              callbacks=[checkpoint])

# Saving themodel
transfer_model.save("model/"+datetime.now().strftime("%Y%m%d-%H%M%S")+"vggforsp500.h5")

#Display the graph of the model
keras.utils.plot_model(transfer_model)

##Display summary of neural network
transfer_model.summary()
